fix(batch): stop the AI usage-count runs overlapping and counting twice
The hourly and overnight forms of ai:usage-counts:update are separate schedule
entries, so Laravel's overlap guard keys them separately and never holds one off
against the other. The overnight recount runs long enough that the next hourly
pass starts during it, then finishes last and writes back the position it read at
the start, moving it backwards. The hourly pass adds to the counts rather than
setting them, so the overlapping range is counted twice until the next recount.

Both forms now take one shared lock, and the stored position only ever moves
forward, which also covers MAX(id) answering from a node that is behind.

Also make AutoRepostService read the repost settings the same way in both places.
The candidate window casts them to int and the per-post decision did not, so a
fractional setting would round differently in each and a post could pass the
second test having already been excluded by the first. Every group stores whole
numbers today; this stops the two halves being able to disagree.

Found by adversarial review.

+++ iznik-batch/app/Services/AutoRepostService.php
@@ -136,12 +136,18 @@ class AutoRepostService
                 continue;
             }
 
+            // Cast the same way applyDueWindow does. The database is asked for candidates
+            // using a window built from whole numbers, and this decides what to do with
+            // them; if the two rounded a fraction differently, a post could pass the test
+            // here that the window had already excluded, and it would never be reposted.
+            // Every community stores whole numbers today, so this changes nothing now -
+            // it just stops the two halves being able to disagree.
             $interval = $msg->type === Message::TYPE_OFFER
-                ? ($reposts['offer'] ?? 3)
-                : ($reposts['wanted'] ?? 7);
+                ? (int) ($reposts['offer'] ?? 3)
+                : (int) ($reposts['wanted'] ?? 7);
 
             // V1: max age check — messages older than interval * (max + 1) days.
-            $maxAge = $interval * (($reposts['max'] ?? 5) + 1);
+            $maxAge = $interval * ((int) ($reposts['max'] ?? 5) + 1);
             if ($msg->hoursago >= $maxAge * 24) {
                 $stats['skipped']++;
                 continue;
